Add lua script to load, merge, and re-serialize dictionary files.
The goal was to have the minimal number and size of diffs when
applied just to the existing dictionary. This succeeds: the order
of words with apostrophes changed, a few minor errors are cleaned up
(unnecessary trailing comma, two lines in a row with the same incorrect
spelling but different correction), but otherwise the output is identical
to the input.

Signed-off-by: Ryan Pavlik <rpavlik@iastate.edu>